Bumping go
Improving mutex usage around writeAtomically. I believe I was missing Unlock in the error case in Set in the unlikely event of any error. This would also hit in redux when writing it - in both cases it should result in deadlock

Handling unlikely event of missing new file when writing atomically. I think this might happen with nil readers, but I couldn't repro this in test. Still - this should be handled properly

Removing legacy code and main package